+++ app/controllers/registrations_controller.rb
@@ -2,6 +2,10 @@ class RegistrationsController < Devise::RegistrationsController
 
   layout 'layouts/application-no-sidebar'
 
+  respond_to :json
+
+  skip_before_action :verify_authenticity_token, only: [ :create ]
+
   protected
 
   def after_sign_up_path_for(resource)